Fixed Auth on DELETE api/lessons/{lesson_id}

+++ src/routes.php
@@ -91,21 +91,20 @@ $app->put('/api/lessons/{lesson_id}', function($req, $res, $args) {
 });
 
 $app->delete('/api/lessons/{lesson_id}', function($req, $res, $args) {
-    //$auth = new \Pond\Auth($this);
+    $auth = new \Pond\Auth($this);
     try{
         $lessons = Pond\Lesson::findOrFail($args['lesson_id']);
-        //$creator_id = $lessons->creator_id;
-        //$isAuth = $auth->isRequestAuthorized($req,$creator_id);
-        //if(!$isAuth) {
-        //    $res->withStatus(401); // Unauthorized
-        //} else {
-            $stat = new \Pond\StatusContainer($lessons);
-            $stat->success();
-            $lessons->delete();
-            $stat->message("The lesson has been deleted");
-            return $res->withJson($stat);
-        //}
+        $creator_id = $lessons->creator_id;
+        if(!$auth->isRequestAuthorized($req,$creator_id)) {
+            return $res->withStatus(401);
+        }
 
+        $stat = new \Pond\StatusContainer($lessons);
+        $stat->success();
+        $lessons->delete();
+        $stat->message("The lesson has been deleted");
+        $res = $res->withStatus(200);
+        return $res->withJson($stat);
     }
     catch(ModelNotFoundException $e){
         $stat = new \Pond\StatusContainer($lessons);
